Removed all eVar69 required fields
I discovered that there is a flaw in the getDaysSinceLastVisit method that eVar69 uses. This function was created by Adobe and we can't change it. Since we can not modify it, I am removing eVar69 required field validations to prevent error noise that we do not have control over. This will be corrected in Web SDK.